[Refactor] cache repository와 postgreSQL repository 분리 & repository 재구조화
