+++ src/data/collectors/landmark_collector.py
@@ -5,8 +5,8 @@ import pandas as pd
 
 from src.infrastructure.external.client.kakao_client import KakaoClient
 from src.data.utils import CollectorUtils
-from src.repository.route.edge_repository import EdgeRepository
-from src.repository.route.poi_repository import LandmarkRepository
+from src.repository.network.edge_repository import EdgeRepository
+from src.repository.layer.landmark_repository import LandmarkRepository
 
 
 class LandmarkCollector:
