+++ src/data/utils/collector_utils.py
@@ -2,8 +2,8 @@ import math
 
 from geoalchemy2.elements import WKTElement
 
-from src.repository.route.edge_repository import EdgeRepository
-from src.repository.route.node_repository import NodeRepository
+from src.repository.network.edge_repository import EdgeRepository
+from src.repository.network.node_repository import NodeRepository
 
 
 class CollectorUtils:
